Refactor transform helper fns; get source input for output transform working
